# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciamento de usuários e tarefas usando Gin e MongoDB",
        "title": "API de Autenticação e Tarefas",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/tasks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Pegar a paginação de tarefas na API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Pegar a paginação",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Número da pagina",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Número de registros por página",
                        "name": "limitPage",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Buscar por qual propriedade de task",
                        "name": "searchField",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Buscar pelo valor da propriedade da task",
                        "name": "searchValue",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Ordenar por qual propriedade de task",
                        "name": "sortField",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ascending",
                            "descending"
                        ],
                        "type": "string",
                        "description": "Ordenação",
                        "name": "sortOrder",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de tasks",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pagination.PaginationResult-entities_Task"
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra uma nova tarefa na API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Criar uma nova tarefa",
                "parameters": [
                    {
                        "description": "Dados do usuário",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.Task"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tarefa criada",
                        "schema": {
                            "$ref": "#/definitions/dtos.Message"
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIError"
                        }
                    },
                    "409": {
                        "description": "Erro de conflito, dados ja existem",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIError"
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Busca uma nova tarefa na API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Buscar uma nova tarefa por ID",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"60c72b2f9b1d8b57b8ed2123\"",
                        "description": "ID da Task",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarefa por ID",
                        "schema": {
                            "$ref": "#/definitions/entities.Task"
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Edita uma tarefa na API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Editar uma tarefa por ID",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"60c72b2f9b1d8b57b8ed2123\"",
                        "description": "ID da Task",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do usuário",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarefa por ID",
                        "schema": {
                            "$ref": "#/definitions/dtos.Message"
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deleta uma tarefa na API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Deletar uma tarefa por ID",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"60c72b2f9b1d8b57b8ed2123\"",
                        "description": "ID da Task",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarefa por ID",
                        "schema": {
                            "$ref": "#/definitions/dtos.Message"
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIError"
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "description": "Registra um novo usuário na API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Criar um novo usuário",
                "parameters": [
                    {
                        "description": "Dados do usuário",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Usuário criado",
                        "schema": {
                            "$ref": "#/definitions/dtos.Message"
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIError"
                        }
                    },
                    "409": {
                        "description": "Erro de conflito, dados ja existem",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIError"
                        }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "description": "Fazer o login do usuário",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Criar o login do usuário",
                "parameters": [
                    {
                        "description": "Dados do usuário",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário login",
                        "schema": {
                            "$ref": "#/definitions/dtos.Token"
                        }
                    },
                    "400": {
                        "description": "Erro de login",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dtos.APIError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "dtos.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "dtos.Task": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dtos.Token": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "dtos.User": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "entities.Task": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "string"
                },
                "createAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updateAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "pagination.PaginationResult-entities_Task": {
            "type": "object",
            "properties": {
                "Err": {},
                "hasNextPage": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Task"
                    }
                },
                "pageCurrent": {
                    "type": "integer"
                },
                "totalItems": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Value: Bearer abc... (Bearer+space+token)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}